+++ x/evm/events.go
@@ -7,19 +7,15 @@ const (
 	EventTypeBlockBloom = "block_bloom"
 	EventTypeTxLog      = "tx_log"
 
-	AttributeKeyContractAddress = "contract"
-	AttributeKeyRecipient       = "recipient"
-	AttributeKeyTxHash          = "txHash"
-	AttributeKeyEthereumTxHash  = "ethereumTxHash"
-	AttributeKeyTxIndex         = "txIndex"
-	AttributeKeyTxGasUsed       = "txGasUsed"
-	AttributeKeyTxType          = "txType"
-	AttributeKeyTxLog           = "txLog"
+	AttributeKeyRecipient      = "recipient"
+	AttributeKeyTxHash         = "txHash"
+	AttributeKeyEthereumTxHash = "ethereumTxHash"
+	AttributeKeyTxIndex        = "txIndex"
+	AttributeKeyTxGasUsed      = "txGasUsed"
+	AttributeKeyTxType         = "txType"
+	AttributeKeyTxLog          = "txLog"
 	// tx failed in eth vm execution
 	AttributeKeyEthereumTxFailed = "ethereumTxFailed"
 	AttributeValueCategory       = ModuleName
 	AttributeKeyEthereumBloom    = "bloom"
-
-	MetricKeyTransitionDB = "transition_db"
-	MetricKeyStaticCall   = "static_call"
 )